libservo: Rename `CookieManager` to `NetworkManager` and move `clear_cookies` to `SiteDataManager`
This patch renames `CookieManager` to `NetworkManager` to better reflect its
future scope. It also moves `clear_cookies` to `SiteDataManager`, aligning
cookie clearing with other planned site data APIs.

No functional changes are intended.
