+++ components/servo/Cargo.toml
@@ -152,9 +152,6 @@ tracing = { workspace = true }
 winit = { workspace = true }
 
 [[test]]
-name = "cookies"
-
-[[test]]
 name = "webview"
 
 [[test]]
