+++ components/servo/tests/site_data.rs
@@ -10,9 +10,52 @@ use http_body_util::combinators::BoxBody;
 use hyper::body::{Bytes, Incoming};
 use hyper::{Request as HyperRequest, Response as HyperResponse};
 use net::test_util::{make_body, make_server};
-use servo::WebViewBuilder;
+use servo::{JSValue, WebViewBuilder};
 
-use crate::common::{ServoTest, WebViewDelegateImpl};
+use crate::common::{ServoTest, WebViewDelegateImpl, evaluate_javascript};
+
+#[test]
+fn test_clear_cookies() {
+    let servo_test = ServoTest::new();
+
+    static MESSAGE: &'static [u8] = b"<!DOCTYPE html>\nHello";
+
+    let handler =
+        move |_: HyperRequest<Incoming>,
+              response: &mut HyperResponse<BoxBody<Bytes, hyper::Error>>| {
+            *response.body_mut() = make_body(MESSAGE.to_vec());
+        };
+    let (server, url) = make_server(handler);
+
+    let delegate = Rc::new(WebViewDelegateImpl::default());
+
+    let webview = WebViewBuilder::new(servo_test.servo(), servo_test.rendering_context.clone())
+        .delegate(delegate.clone())
+        .url(url.into_url())
+        .build();
+
+    servo_test.spin(move || !delegate.url_changed.get());
+
+    let _ = server.close();
+
+    let result = evaluate_javascript(
+        &servo_test,
+        webview.clone(),
+        "document.cookie = 'foo1=bar1';\
+         document.cookie = 'foo2=bar2';\
+         document.cookie = 'foo3=bar3';\
+         document.cookie;",
+    );
+    assert_eq!(
+        result,
+        Ok(JSValue::String("foo1=bar1; foo2=bar2; foo3=bar3".into()))
+    );
+
+    servo_test.servo().site_data_manager().clear_cookies();
+
+    let result = evaluate_javascript(&servo_test, webview.clone(), "document.cookie");
+    assert_eq!(result, Ok(JSValue::String("".into())));
+}
 
 #[test]
 fn test_clear_cache() {
