+++ ports/servoshell/webdriver.rs
@@ -139,7 +139,7 @@ impl RunningAppState {
         while let Ok(msg) = webdriver_receiver.try_recv() {
             match msg {
                 WebDriverCommandMsg::ResetAllCookies(sender) => {
-                    self.servo().cookie_manager().clear_cookies();
+                    self.servo().site_data_manager().clear_cookies();
                     let _ = sender.send(());
                 },
                 WebDriverCommandMsg::Shutdown => {
